chore: Update npm dependencies and add @urql/core version 4.3.0
This commit updates the npm dependencies in the package.json file. Specifically, it updates the version of "@rahataid/c2c-query" to "0.0.5" and adds the new dependency "@urql/core" with version 4.3.0. These updates ensure that the project is using the latest stable versions of these packages, which may include bug fixes, performance improvements, and new features.