chore: Update beneficiary table component
This commit updates the `beneficiary.table.tsx` file in the `rahat-ui` repository. It adds the `useRouter` hook from `next/navigation` to handle routing. This change aims to improve the functionality of the beneficiary table component by allowing users to navigate to the "Add Beneficiary Data" page when clicking the "Add" button.